Start Object refactoring
