+++ app/object_values/base.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class ObjectValue(BaseModel):
+
+    class Config:
+        allow_mutation = False
